CE 631: Refactor file structure (#490)
