+++ apps/cave/__tests__/utils/formatFixed.test.ts
@@ -1,5 +1,5 @@
 import { BigNumber } from 'ethers'
-import { compactFormat, formatFixed } from '../../utils/formatFixed'
+import { compactFormat, formatFixed } from '../../utils/bigNumberMask'
 
 describe('test numberMask', () => {
   it('test with string', async () => {
