+++ apps/cave/components/AMM/AddLiquidity/AddLiquidity.tsx
@@ -4,19 +4,21 @@ import { PlusIcon } from '@concave/icons'
 import { Button, ButtonProps, Card, Flex, Modal, Text, useDisclosure } from '@concave/ui'
 import { CurrencyInputField } from 'components/AMM'
 import { SupplyLiquidityModal } from 'components/AMM/AddLiquidity/SupplyLiquidityModal'
-import { useAddLiquidityButtonProps } from 'components/AMM/AddLiquidity/useAddLiquidityButtonProps'
-import { useAddLiquidityState } from 'components/AMM/AddLiquidity/useAddLiquidityState'
-import { useAddLiquidityTransaction } from 'components/AMM/AddLiquidity/useAddLiquidityTransaction'
-import { ConnectButton } from 'components/ConnectWallet'
+import { useAddLiquidityButtonProps } from 'components/AMM/hooks/useAddLiquidityButtonProps'
+
+import { ConnectButton } from 'components/YourWallet/ConnectWallet'
 import { SelectAMMCurrency } from 'components/CurrencySelector/SelectAMMCurrency'
-import { TransactionErrorDialog } from 'components/TransactionErrorDialog'
-import { TransactionSubmittedDialog } from 'components/TransactionSubmittedDialog'
-import { WaitingConfirmationDialog } from 'components/WaitingConfirmationDialog'
+import { TransactionErrorDialog } from 'components/TransactionDialog/TransactionErrorDialog'
+import { TransactionSubmittedDialog } from 'components/TransactionDialog/TransactionSubmittedDialog'
+
+import { WaitingConfirmationDialog } from 'components/TransactionDialog/TransactionWaitingConfirmationDialog'
 import { toAmount } from 'utils/toAmount'
 import { useAccount } from 'wagmi'
+import { useAddLiquidityState } from '../hooks/useAddLiquidityState'
+import { useAddLiquidityTransaction } from '../hooks/useAddLiquidityTransaction'
+import useLiquidityData from '../hooks/useLiquidityData'
 import { useQueryCurrencies } from '../hooks/useQueryCurrencies'
 import { NetworkMismatch } from '../NetworkMismatch'
-import useLiquidityData from './useLiquidityData'
 
 const AddSymbol = () => (
   <Flex align="center" justify="center">
