+++ apps/cave/components/AMM/hooks/useSwapTransaction.ts
@@ -8,6 +8,7 @@ import { toPercent } from 'utils/toPercent'
 import { useAccount, useContractWrite, useNetwork } from 'wagmi'
 import { useSwapSettings } from '../Swap/Settings'
 
+
 export const useSwapTransaction = (
   _trade: Trade<Currency, Currency, TradeType>,
   recipient: string,
