+++ apps/cave/components/AMM/index.ts
@@ -1,6 +1,9 @@
 export { CurrencyInputField } from '../CurrencyAmountField'
 export { CandleStickCard } from './Chart'
 export { GasPrice } from './GasPrice'
+export { useSwapButtonProps } from './hooks/useSwapButtonProps'
+export { useSwapState } from './hooks/useSwapState'
+export { useSwapTransaction } from './hooks/useSwapTransaction'
 export { ConfirmSwapModal } from './Swap/ConfirmSwapModal'
 export { CurrencyOutputField } from './Swap/CurrencyOutputField'
 export { CustomRecipient } from './Swap/CustomRecipient'
@@ -8,6 +11,3 @@ export { RelativePrice } from './Swap/RelativePrice'
 export { Settings } from './Swap/Settings'
 export type { SwapSettings } from './Swap/Settings'
 export { SwitchCurrencies } from './Swap/SwitchCurrencies'
-export { useSwapButtonProps } from './Swap/useSwapButtonProps'
-export { useSwapState } from './Swap/useSwapState'
-export { useSwapTransaction } from './Swap/useSwapTransaction'
