+++ apps/cave/components/CurrencyAmountField/index.tsx
@@ -1,3 +1,3 @@
 export { CurrencyAmountField } from './CurrencyAmountField'
 export { CurrencyInputField } from './CurrencyInputField'
-export { useLinkedCurrencyFields } from './useLinkedCurrencyFields'
+export { useLinkedCurrencyFields } from './hooks/useLinkedCurrencyFields'
