+++ apps/cave/components/Marketplace/Main/MarketplacePosition.tsx
@@ -5,13 +5,13 @@ import { FixedOrderMarketContract, StakingPosition } from '@concave/marketplace'
 import { FlexProps, gradientBorder, HStack, Spinner } from '@concave/ui'
 import { BuyButton } from 'components/BuyButton/BuyButton'
 import { useCurrencyButtonState } from 'components/CurrencyAmountButton/CurrencyAmountButton'
-import { usePositionDiscount } from 'components/StakingPositions/LockPosition/MarketLockInfo/usePositionDiscount'
+import { usePositionDiscount } from 'components/StakingPositions/LockPosition/MarketLockInfo/hooks/usePositionDiscount'
 import { differenceInDays, format, formatDistanceToNowStrict } from 'date-fns'
 import { useTransaction } from 'hooks/TransactionsRegistry/useTransaction'
 import { useCurrentSupportedNetworkId } from 'hooks/useCurrentSupportedNetworkId'
 import { concaveProvider } from 'lib/providers'
 import { useMemo, useState } from 'react'
-import { compactFormat, formatFixed } from 'utils/formatFixed'
+import { compactFormat, formatFixed } from 'utils/bigNumberMask'
 import { useAccount, useSigner } from 'wagmi'
 import { ConfirmPurchaseModal } from './ConfirmBuy'
