+++ apps/cave/components/SideBar/SideBarTop.tsx
@@ -1,8 +1,8 @@
 import { CNV } from '@concave/core'
 import { DashboardIcon } from '@concave/icons'
 import { Box, Flex, Image, Stack, Text } from '@concave/ui'
-import { ButtonLink } from 'components/ButtonLink'
-import { ConnectButton, UserWallet } from 'components/ConnectWallet'
+import { ButtonLink } from 'components/SideBar/ButtonLink'
+import { ConnectButton, UserWallet } from 'components/YourWallet/ConnectWallet'
 import { useCurrencyBalance } from 'hooks/useCurrencyBalance'
 import { useCurrentSupportedNetworkId } from 'hooks/useCurrentSupportedNetworkId'
 import { useAccount, useNetwork } from 'wagmi'
